requirements.txt
# Core - OpenVINO inference
openvino>=2025.1
openvino-genai>=2025.1

# Model export
optimum[openvino]
transformers

# Parakeet TDT preprocessing
onnxruntime>=1.18.1

# Audio
sounddevice
numpy

# Input/output
keyboard
pyperclip

# GUI
pystray>=0.19
customtkinter>=5.2
Pillow>=10.0

# Testing
pytest